Refactor: import path name change
