+++ src/components/cards/coinCard/price/PriceChange.tsx
@@ -1,6 +1,6 @@
 import { memo } from "react";
 import useColorChange from "use-color-change";
-import { currencyValueFormat } from "@/helpers/numberFromat";
+import { currencyValueFormat } from "src/utils/numberFromat";
 
 const Price = ({ price }: { price: number }) => {
   const colorStyle = useColorChange(price, {
