+++ src/components/currencies/currencyDetail/coinStatistic/CoinStatistic.tsx
@@ -1,4 +1,4 @@
-import { currencyValueFormat } from "@/helpers/numberFromat";
+import { currencyValueFormat } from "src/utils/numberFromat";
 import { ICoin } from "@/types/types";
 import PriceTimeChange from "../../../cards/coinCard/price/PriceTimeChange";
 import classes from "./CoinStatistic.module.scss";
